+++ vnext/include/Include.vcxitems
@@ -17,6 +17,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)Shared\AsyncStorageModuleWin32Config.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Shared\cdebug.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Shared\II18nModule.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)Shared\magic_enum.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Shared\ReactWindowsAPI.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Shared\ViewManager.h" />
   </ItemGroup>
