+++ vnext/include/Shared/magic_enum.hpp
